requirements.txt
flask
PyMySQL
sqlalchemy
redis
dogpile.cache
kafka-python
flask-socketio
flask_cors
cryptography
